feat: 병합후 위치조정
