--- src/main/java/org/example/anonymous/domain/product/dto/request/ChangeQuantityRequest.java
@@ -1,5 +1,0 @@
-package org.example.anonymous.domain.product.dto.request;
-
-public record ChangeQuantityRequest(int quantity) {
-
-}
