--- src/main/java/org/example/anonymous/domain/product/dto/request/ProductRequestDto.java
@@ -1,16 +1,0 @@
-package org.example.anonymous.domain.product.dto.request;
-
-public class ProductRequestDto {
-    private String name;
-
-    public ProductRequestDto() {
-    }
-
-    public ProductRequestDto(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
