--- src/main/java/org/example/anonymous/domain/product/dto/response/ProductResponseDto.java
@@ -1,20 +1,0 @@
-package org.example.anonymous.domain.product.dto.response;
-
-public class ProductResponseDto {
-
-    private Long productId;
-    private String name;
-
-    public ProductResponseDto(Long productId, String name) {
-        this.productId = productId;
-        this.name = name;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
