--- src/main/java/org/example/anonymous/domain/product/entity/Cart.java
@@ -1,29 +1,0 @@
-package org.example.anonymous.domain.product.entity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.example.anonymous.common.base.BaseEntity;
-
-@Entity
-@Table(name = "cart")
-@NoArgsConstructor
-@Getter
-public class Cart extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<CartItem> cartItems = new ArrayList<>();
-
-}
