--- src/main/java/org/example/anonymous/domain/product/entity/CartItem.java
@@ -1,36 +1,0 @@
-package org.example.anonymous.domain.product.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Entity
-@Table(name = "cart_item")
-@Getter
-public class CartItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private int quantity;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "cart_id")
-    private Cart cart;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id")
-    private Product product;
-
-    public CartItem(int quantity, Cart cart, Product product) {
-        this.quantity = quantity;
-        this.cart = cart;
-        this.product = product;
-    }
-
-    public void updateQuantity(int quantity){
-        this.quantity = quantity;
-    }
-}
