--- src/main/java/org/example/anonymous/domain/product/entity/ProductStatus.java
@@ -1,5 +1,0 @@
-package org.example.anonymous.domain.product.entity;
-
-public enum ProductStatus {
-    FOR_SALE,SOLD_OUT
-}
