--- src/main/java/org/example/anonymous/domain/product/repository/CartRepository.java
@@ -1,9 +1,0 @@
-package org.example.anonymous.domain.product.repository;
-
-import org.example.anonymous.domain.product.entity.Cart;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CartRepository extends JpaRepository<Cart,Long> {
-}
