--- src/main/java/org/example/anonymous/domain/product/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package org.example.anonymous.domain.product.repository;
-
-import org.example.anonymous.domain.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
