+++ src/main/java/org/example/hansabal/domain/product/dto/request/ChangeQuantityRequest.java
@@ -1,0 +1,5 @@
+package org.example.hansabal.domain.product.dto.request;
+
+public record ChangeQuantityRequest(int quantity) {
+
+}
