+++ src/main/java/org/example/hansabal/domain/product/dto/response/ItemResponse.java
@@ -1,0 +1,12 @@
+package org.example.hansabal.domain.product.dto.response;
+
+import org.example.hansabal.domain.product.entity.CartItem;
+
+public record ItemResponse(
+        String ProductName,
+        int quantity
+) {
+    public static ItemResponse from(CartItem cartItem){
+        return new ItemResponse(cartItem.getProduct().getName(),cartItem.getQuantity());
+    }
+}
