+++ src/main/java/org/example/hansabal/domain/product/repository/CartRepository.java
@@ -1,0 +1,9 @@
+package org.example.hansabal.domain.product.repository;
+
+import org.example.hansabal.domain.product.entity.Cart;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CartRepository extends JpaRepository<Cart,Long> {
+}
