+++ src/main/java/org/example/hansabal/product/controller/ProductController.java
@@ -1,0 +1,45 @@
+package org.example.hansabal.product.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.example.hansabal.product.dto.request.ProductRequestDto;
+import org.example.hansabal.product.dto.response.ProductResponseDto;
+import org.example.hansabal.product.service.ProductService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/product")
+public class ProductController {
+    private final ProductService productService;
+
+    @PostMapping
+    public ResponseEntity<ProductResponseDto> createProduct(@RequestBody ProductRequestDto requestDto) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(productService.createProduct(requestDto));
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<ProductResponseDto> getProduct(@PathVariable Long id) {
+        return ResponseEntity.ok(productService.getProduct(id));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ProductResponseDto>> getAllProducts() {
+        return ResponseEntity.ok(productService.getAllProducts());
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<ProductResponseDto> updateProduct(@PathVariable Long id,
+                                                            @RequestBody ProductRequestDto requestDto) {
+        return ResponseEntity.ok(productService.updateProduct(id, requestDto));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        productService.deleteProduct(id);
+        return ResponseEntity.noContent().build();
+    }
+}
