+++ src/main/java/org/example/hansabal/product/dto/request/ChangeQuantityRequest.java
@@ -1,0 +1,5 @@
+package org.example.hansabal.product.dto.request;
+
+public record ChangeQuantityRequest(int quantity) {
+
+}
