+++ src/main/java/org/example/hansabal/product/dto/request/ProductRequestDto.java
@@ -1,0 +1,16 @@
+package org.example.hansabal.product.dto.request;
+
+public class ProductRequestDto {
+    private String name;
+
+    public ProductRequestDto() {
+    }
+
+    public ProductRequestDto(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
