+++ src/main/java/org/example/hansabal/product/entity/ProductStatus.java
@@ -1,0 +1,5 @@
+package org.example.hansabal.product.entity;
+
+public enum ProductStatus {
+    FOR_SALE,SOLD_OUT
+}
