+++ src/main/java/org/example/hansabal/product/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package org.example.hansabal.product.repository;
+
+import org.example.hansabal.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
